SPP seems to work fine in parallel. I have one last change to do with letting the command line argument for #threads work. Also, i forgot to change the inverse functions in each diagram to make use of the direct channel index matrix, which meant they used WAY more time than necessary. Think I'm running out of things to fix, finally

+++ CCD_prototype/CCD/makeampmat.h
@@ -44,7 +44,7 @@ public:
     //std::vector<double>                       T2_elements_A;
     //std::map<unsigned long int, int> T3_elements_I_um;
     //std::map<unsigned long int, int>        T3_elements_I;        //holds indices to T3_elements_A, same for _new and _temp
-    spp::sparse_hash_map<unsigned long int, int>        T3_elements_I;
+    spp::sparse_hash_map<unsigned long int, unsigned long int>        T3_elements_I;
     //std::vector<spp::sparse_hash_map<unsigned long int, int>> T3_elements_IV;
     std::vector<double>                       T3_elements_A;        //holds T3 amplitudes
     std::vector<double>                       T3_elements_A_new;    //holds new T3 amplitudes
